Updates API V0.1 documentation  (#52)
Summary: This PR updates the API V0.1 documentation and example generation script, adding new features such as the ARNI rir db and velocity control for moving sound events. It enhances the documentation with updates to the readme, script formatting improvements, the addition of a license shield and development warning, and more detailed and formatted docstrings. Additionally, it addresses an edge case for n_events < 1 and includes black formatting and bibtex updates to ensure consistency and comprehensive guidance for users.